+++ include/engine/scenes.h
@@ -7,13 +7,13 @@
 #define Amphora_BeginScene(name) \
 extern "C" { \
 extern void name##_Init(void); \
-extern void name##_Update(Uint64, const InputState *); \
+extern void name##_Update(Uint32, const InputState *); \
 extern void name##_Destroy(void); \
 };
 #else
 #define Amphora_BeginScene(name) \
 extern void name##_Init(void); \
-extern void name##_Update(Uint64, const InputState *); \
+extern void name##_Update(Uint32, const InputState *); \
 extern void name##_Destroy(void);
 #endif
